backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
redis==5.0.4
pydantic==2.7.0
python-dotenv==1.0.1
openai==1.30.0
google-generativeai==0.5.4
spacy==3.7.4
scikit-learn==1.5.0
httpx==0.27.0
python-multipart==0.0.9
llama-cpp-python==0.2.75
